# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TriggeryOmi Drones and Parts API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  imageUrl: string;
  stock: number;
  featured?: boolean;
  /** @nullable */
  averageRating?: number | null;
  reviewCount?: number;
  soldLastMonth?: number;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  price: number;
  category: string;
  imageUrl: string;
  /** @minimum 0 */
  stock: number;
  featured?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  imageUrl?: string;
  stock?: number;
  featured?: boolean;
}

export interface Review {
  id: number;
  productId: number;
  reviewerName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface ReviewInput {
  /** @minLength 1 */
  reviewerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minLength 1 */
  comment: string;
}

export interface OrderItem {
  productId: number;
  quantity: number;
  priceAtPurchase: number;
  productName?: string;
}

export interface CartItem {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  cardNumber?: string;
  cardExpiry?: string;
  cardCvv?: string;
  items: CartItem[];
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  totalAmount: number;
  status: string;
  createdAt: string;
  items: OrderItem[];
}

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  ordersThisMonth: number;
  revenueThisMonth: number;
  /** @nullable */
  topProduct: string | null;
}

export interface MonthlySales {
  month: string;
  orders: number;
  revenue: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
};