# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TriggeryOmi Drones and Parts API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: reviews
    description: Review operations
  - name: orders
    description: Order operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products, admin]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products, admin]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products, admin]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createProductReview
      tags: [reviews]
      summary: Add a review to a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (simulated checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats (sales, orders, revenue)
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders with product details
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/monthly-sales:
    get:
      operationId: getMonthlySales
      tags: [admin]
      summary: Get monthly sales breakdown for charts
      responses:
        "200":
          description: Monthly sales data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlySales"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Product:
      type: object
      required: [id, name, description, price, category, imageUrl, stock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        featured:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        soldLastMonth:
          type: integer
        createdAt:
          type: string

    ProductInput:
      type: object
      required: [name, description, price, category, imageUrl, stock]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
        category:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
          minimum: 0
        featured:
          type: boolean

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        featured:
          type: boolean

    Review:
      type: object
      required: [id, productId, reviewerName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string

    ReviewInput:
      type: object
      required: [reviewerName, rating, comment]
      properties:
        reviewerName:
          type: string
          minLength: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 1

    OrderItem:
      type: object
      required: [productId, quantity, priceAtPurchase]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        priceAtPurchase:
          type: number
        productName:
          type: string

    OrderInput:
      type: object
      required: [customerName, customerEmail, items, shippingAddress]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        cardNumber:
          type: string
        cardExpiry:
          type: string
        cardCvv:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"

    CartItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer

    Order:
      type: object
      required: [id, customerName, customerEmail, shippingAddress, totalAmount, status, createdAt, items]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        totalAmount:
          type: number
        status:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    AdminStats:
      type: object
      required: [totalRevenue, totalOrders, totalProducts, ordersThisMonth, revenueThisMonth, topProduct]
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        totalProducts:
          type: integer
        ordersThisMonth:
          type: integer
        revenueThisMonth:
          type: number
        topProduct:
          type: ["string", "null"]

    MonthlySales:
      type: object
      required: [month, orders, revenue]
      properties:
        month:
          type: string
        orders:
          type: integer
        revenue:
          type: number