# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TriggeryOmi Drones and Parts API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  ordersThisMonth: number;
  revenueThisMonth: number;
  /** @nullable */
  topProduct: string | null;
}