# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TriggeryOmi Drones and Parts API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  cardNumber?: string;
  cardExpiry?: string;
  cardCvv?: string;
  items: CartItem[];
}