# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TriggeryOmi Drones and Parts API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  imageUrl: string;
  stock: number;
  featured?: boolean;
  /** @nullable */
  averageRating?: number | null;
  reviewCount?: number;
  soldLastMonth?: number;
  createdAt: string;
}