# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TriggeryOmi Drones and Parts API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  price: number;
  category: string;
  imageUrl: string;
  /** @minimum 0 */
  stock: number;
  featured?: boolean;
}